requirements.txt
discord.py==2.3.2
ffmpeg_python==0.2.0
